+++ src/sorts/ISorter.java
@@ -4,5 +4,5 @@ public interface ISorter<T extends Comparable<T>> {
     T[] sort(T[] dados);
     int getComparisons();
     int getSwaps();
-    int getTime();
+    long getTime();
 }
